refactor(capture): 重构拍照逻辑并优化 HDR 功能
- 将 MediaSaver 重命名为 CaptureJob， CaptureTaskManager 重命名为 CaptureJobManager
- 优化了拍照流程，支持 HDR拍照功能
- 调整了 ImageProcessor 的 capture 方法参数类型
- 在 MainActivity 中增加了 isCapturing 标志位，用于防止重复拍照
- 修改了拍照相关回调和处理逻辑，以适应新的 HDR 拍照流程

+++ app/src/main/java/com/tainzhi/android/tcamera/ImageProcessor.kt
@@ -12,7 +12,7 @@ object ImageProcessor {
         init(App.getCachePath())
     }
 
-    fun capture(captureType: Int, jobId:Int, timeStamp: String, exposureTimes: List<Int>) {
+    fun capture(captureType: Int, jobId:Int, timeStamp: String, exposureTimes: List<Long>) {
         Log.d(TAG, "capture, jobId:$jobId, timeStamp:$timeStamp")
         val frameSize = 3
         capture(captureType, jobId, timeStamp, frameSize, exposureTimes)
@@ -44,7 +44,7 @@ object ImageProcessor {
     private external fun init(cachePath: String)
     private external fun handlePreviewImage(yPlane: ByteBuffer, uPlane: ByteBuffer, vPlane: ByteBuffer,  width: Int, height: Int)
 
-    private external fun capture(captureType: Int, jobId:Int, timeStamp: String, frameSize: Int, exposureTimes: List<Int>)
+    private external fun capture(captureType: Int, jobId:Int, timeStamp: String, frameSize: Int, exposureTimes: List<Long>)
     private external fun processImage(jobId: Int, yPlane: ByteBuffer, uPlane: ByteBuffer, vPlane: ByteBuffer,  width: Int, height: Int)
     private external fun updateCaptureBackupFilePath(path: String)
     private external fun deinit()
